VZ-8138: Remove duplicated code
