+++ tests/e2e/backup/helpers/models.go
@@ -638,4 +638,5 @@ var (
 	OciCliFingerprint           string
 	OciCliKeyFile               string
 	KeyCloakReplicaCount        int32
+	MySQLBackupMode             string
 )
