+++ tests/e2e/backup/mysql/mysql_backup_suite_test.go
@@ -1,7 +1,7 @@
 // Copyright (c) 2022, 2023, Oracle and/or its affiliates.
 // Licensed under the Universal Permissive License v 1.0 as shown at https://oss.oracle.com/licenses/upl.
 
-package mysqloci
+package mysql
 
 import (
 	"testing"
