+++ tests/e2e/backup/mysql/mysql_backup_test.go
@@ -1,7 +1,7 @@
 // Copyright (c) 2022, 2023, Oracle and/or its affiliates.
 // Licensed under the Universal Permissive License v 1.0 as shown at https://oss.oracle.com/licenses/upl.
 
-package mysqloci
+package mysql
 
 import (
 	"bytes"
@@ -90,6 +90,31 @@ func CreateInnoDBBackupObjectWithOci() error {
 	return nil
 }
 
+// func CreateInnoDBBackupObjectWithS3() creates mysql operator backup resource to start the backup.
+func CreateInnoDBBackupObjectWithS3() error {
+	var b bytes.Buffer
+	template, _ := template.New("mysql-backup").Parse(common.InnoDBBackupS3)
+	data := common.InnoDBBackupObject{
+		InnoDBBackupName:                  common.BackupMySQLName,
+		InnoDBNamespaceName:               constants.KeycloakNamespace,
+		InnoDBClusterName:                 common.InnoDBClusterName,
+		InnoDBBackupProfileName:           common.BackupResourceName,
+		InnoDBBackupObjectStoreBucketName: common.OciBucketName,
+		InnoDBBackupCredentialsName:       common.VeleroMySQLSecretName,
+		InnoDBBackupStorageName:           common.BackupMySQLStorageName,
+		InnoDBObjectStorageNamespaceName:  common.OciNamespaceName,
+		InnoDBBackupRegion:                common.BackupRegion,
+	}
+	template.Execute(&b, data)
+	err := common.DynamicSSA(context.TODO(), b.String(), t.Logs)
+	if err != nil {
+		t.Logs.Errorf("Error creating innodb backup object", zap.Error(err))
+		return err
+	}
+
+	return nil
+}
+
 func BackupMySQLValues() error {
 	t.Logs.Infof("Backing up mysql values to file '%s'", common.MySQLBackupHelmFileName)
 	var cmd common.BashCommand
@@ -155,14 +180,27 @@ func MySQLRestore() error {
 
 	var cmd common.BashCommand
 	var cmdArgs []string
-	cmdArgs = append(cmdArgs, "helm", "install", mysqlChartName, vzMySQLChartPath)
-	cmdArgs = append(cmdArgs, "--namespace", constants.KeycloakNamespace)
-	cmdArgs = append(cmdArgs, "--set", "initDB.dump.name=alpha")
-	cmdArgs = append(cmdArgs, "--set", fmt.Sprintf("initDB.dump.ociObjectStorage.prefix=%s/%s", common.BackupMySQLStorageName, backupFolderName))
-	cmdArgs = append(cmdArgs, "--set", fmt.Sprintf("initDB.dump.ociObjectStorage.bucketName=%s", common.OciBucketName))
-	cmdArgs = append(cmdArgs, "--set", fmt.Sprintf("initDB.dump.ociObjectStorage.credentials=%s", common.VeleroMySQLSecretName))
-	cmdArgs = append(cmdArgs, "--values", common.MySQLBackupHelmFileName)
 
+	if common.MySQLBackupMode == "s3" {
+		s3EndPoint := fmt.Sprintf("https://%s.compat.objectstorage.%s.oraclecloud.com", common.OciNamespaceName, common.BackupRegion)
+		cmdArgs = append(cmdArgs, "helm", "install", mysqlChartName, vzMySQLChartPath)
+		cmdArgs = append(cmdArgs, "--namespace", constants.KeycloakNamespace)
+		cmdArgs = append(cmdArgs, "--set", "initDB.dump.name=alpha")
+		cmdArgs = append(cmdArgs, "--set", fmt.Sprintf("initDB.dump.s3.prefix=%s/%s", common.BackupMySQLStorageName, backupFolderName))
+		cmdArgs = append(cmdArgs, "--set", fmt.Sprintf("initDB.dump.s3.bucketName=%s", common.OciBucketName))
+		cmdArgs = append(cmdArgs, "--set", fmt.Sprintf("initDB.dump.s3.config=%s", common.VeleroMySQLSecretName))
+		cmdArgs = append(cmdArgs, "--set", fmt.Sprintf("initDB.dump.s3.endpoint=%s", s3EndPoint))
+		cmdArgs = append(cmdArgs, "--set", "initDB.dump.s3.profile=default")
+		cmdArgs = append(cmdArgs, "--values", common.MySQLBackupHelmFileName)
+	} else {
+		cmdArgs = append(cmdArgs, "helm", "install", mysqlChartName, vzMySQLChartPath)
+		cmdArgs = append(cmdArgs, "--namespace", constants.KeycloakNamespace)
+		cmdArgs = append(cmdArgs, "--set", "initDB.dump.name=alpha")
+		cmdArgs = append(cmdArgs, "--set", fmt.Sprintf("initDB.dump.ociObjectStorage.prefix=%s/%s", common.BackupMySQLStorageName, backupFolderName))
+		cmdArgs = append(cmdArgs, "--set", fmt.Sprintf("initDB.dump.ociObjectStorage.bucketName=%s", common.OciBucketName))
+		cmdArgs = append(cmdArgs, "--set", fmt.Sprintf("initDB.dump.ociObjectStorage.credentials=%s", common.VeleroMySQLSecretName))
+		cmdArgs = append(cmdArgs, "--values", common.MySQLBackupHelmFileName)
+	}
 	cmd.CommandArgs = cmdArgs
 
 	response := common.Runner(&cmd, t.Logs)
@@ -342,9 +380,15 @@ func backupPrerequisites() {
 		return BackupMySQLValues()
 	}, shortWaitTimeout, shortPollingInterval).Should(BeNil())
 
-	Eventually(func() error {
-		return common.CreateMySQLCredentialsSecretFromUserPrincipal(constants.KeycloakNamespace, common.VeleroMySQLSecretName, t.Logs)
-	}, shortWaitTimeout, shortPollingInterval).Should(BeNil())
+	if common.MySQLBackupMode == "s3" {
+		Eventually(func() error {
+			return common.CreateMySQLCredentialsSecretFromFile(constants.KeycloakNamespace, common.VeleroMySQLSecretName, t.Logs)
+		}, shortWaitTimeout, shortPollingInterval).Should(BeNil())
+	} else {
+		Eventually(func() error {
+			return common.CreateMySQLCredentialsSecretFromUserPrincipal(constants.KeycloakNamespace, common.VeleroMySQLSecretName, t.Logs)
+		}, shortWaitTimeout, shortPollingInterval).Should(BeNil())
+	}
 
 	t.Logs.Info("Create a sample keycloak user")
 	Eventually(func() error {
